update from dev

+++ app/Http/Controllers/Api/KeywordController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Api;
 
 use App\Key;
+use App\Module;
 use App\Keyword;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
@@ -10,8 +11,17 @@ use App\Http\Requests\KeywordRequest;
 
 class KeywordController extends Controller
 {
+    protected $module;
+
+    public function __construct()
+    {
+        $this->module = Module::where('code', 'keyword')->first();
+    }
+
     public function index()
     {
+        $this->authorize('view', $this->module);
+
         return Key::with('keywords')->get();
     }
 
@@ -22,32 +32,38 @@ class KeywordController extends Controller
 
     public function store(KeywordRequest $request)
     {
+        $this->authorize('create', $this->module);
+
         $key = Key::where('id', request('key'))->first();
 
         $keyword = $key->keywords()->create($request->only('value'));
 
         return response()->json([
-            'message' => 'Successfuly created key keyword.',
+            'message' => 'Successfuly created keyword.',
             'keyword' => $keyword
         ]);
     }
 
     public function update(KeywordRequest $request, Keyword $keyword)
     {
+        $this->authorize('update', $this->module);
+
         $keyword->update($request->only('value'));
 
         return response()->json([
-            'message' => 'Successfully updated key keyword.',
+            'message' => 'Successfully updated keyword.',
             'keyword' => $keyword
         ]);
     }
 
     public function destroy(Keyword $keyword)
     {
+        $this->authorize('destroy', $this->module);
+
         $keyword->delete();
 
         return [
-            'message' => 'Successfully deleted key keyword.'
+            'message' => 'Successfully deleted keyword.'
         ];
     }
 }
